+++ frontend/src/Commons/Forms/InputModules/index.ts
@@ -3,4 +3,3 @@ export { default as FieldsAndSoil } from './FieldsAndSoil/FieldsAndSoil.tsx';
 export { default as Crops } from './Crops/CropsInfo.tsx';
 export { default as Fertilizers } from './Fertilizers/Fertilizers.tsx';
 export { default as CalculateNutrients } from './CalculateNutrients/CalculateNutrients.tsx';
-export { default as Summary } from './Summary/Summary.tsx';
