+++ frontend/src/Constants/ModuleIDs.ts
@@ -2,5 +2,4 @@ export const FARM_INFORMATION = 'FarmInformation';
 export const FIELDS_AND_SOIL = 'FieldsAndSoil';
 export const CROPS_INFORMATION = 'Crops';
 export const FERTILIZERS_INFORMATION = 'Fertilizers';
-export const CALCULATION_INFORMATION = 'Calculation';
-export const SUMMARY = 'Summary';
+export const CALCULATE_NUTRIENTS = 'CalculateNutrients';
