+++ frontend/src/Constants/Names.ts
@@ -1,6 +1,8 @@
 const Names = {
   FARM_DETAILS: 'farmDetails',
   FERTILIZER_DETAILS: 'fertilizersDetails',
+  FORM_STATES: 'formStates',
+  CURRENT_FORM: 'currForm',
 };
 
 export default Names;
